# docs/api/venice_openapi.yaml
...
info:
    description: The Venice.ai API.
    termsOfService: https://venice.ai/legal/tos
    title: Venice.ai API
    version: "20250805.225500"
openapi: 3.0.0
security:
    - BearerAuth: []
servers:
    - url: https://api.venice.ai/api/v1
...
components:
    securitySchemes:
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http
    schemas:
        StandardError:
            type: object
            properties:
                error:
                    type: string
                    description: A description of the error
            required:
                - error
        DetailedError:
            type: object
            properties:
                details:
                    type: object
                    properties: {}
                    description: Details about the incorrect input
                    example:
                        _errors: []
                        field:
                            _errors:
                                - Field is required
                error:
                    type: string
                    description: A description of the error
            required:
                - error
...
        SimpleGenerateImageRequest:
            type: object
            properties:
                background:
                    type: string
                    nullable: true
                    enum:
                        - transparent
                        - opaque
                        - auto
                    default: auto
                    description: This parameter is not used in Venice image generation but is
                        supported for compatibility with OpenAI API
                    example: auto
                model:
                    type: string
                    default: default
                    description: The model to use for image generation. Defaults to Venice's default
                        image model. If a non-existent model is specified (ie an OpenAI
                        model name), it will default to Venice's default image model.
                    example: hidream
                moderation:
                    type: string
                    nullable: true
                    enum:
                        - low
                        - auto
                    default: auto
                    description: auto enables safe venice mode which will blur out adult content.
                        low disables safe venice mode.
                    example: auto
                n:
                    type: integer
                    nullable: true
                    minimum: 1
                    maximum: 1
                    default: 1
                    description: Number of images to generate. Venice presently only supports 1
                        image per request.
                    example: 1
                output_compression:
                    type: integer
                    nullable: true
                    minimum: 0
                    maximum: 100
                    default: 100
                    description: This parameter is not used in Venice image generation but is
                        supported for compatibility with OpenAI API
                output_format:
                    type: string
                    enum:
                        - jpeg
                        - png
                        - webp
                    default: png
                    description: Output format for generated images
                    example: png
                prompt:
                    type: string
                    minLength: 1
                    maxLength: 1500
                    description: A text description of the desired image.
                    example: A beautiful sunset over mountain ranges
                quality:
                    type: string
                    nullable: true
                    enum:
                        - auto
                        - high
                        - medium
                        - low
                        - hd
                        - standard
                    default: auto
                    description: This parameter is not used in Venice image generation but is
                        supported for compatibility with OpenAI API
                    example: auto
                response_format:
                    type: string
                    nullable: true
                    enum:
                        - b64_json
                        - url
                    default: b64_json
                    description: Response format. URL will be a data URL.
                    example: b64_json
                size:
                    type: string
                    nullable: true
                    enum:
                        - auto
                        - 256x256
                        - 512x512
                        - 1024x1024
                        - 1536x1024
                        - 1024x1536
                        - 1792x1024
                        - 1024x1792
                    default: auto
                    description: Size of generated images. Default is 1024x1024
                    example: 1024x1024
                style:
                    type: string
                    nullable: true
                    enum:
                        - vivid
                        - natural
                    default: natural
                    description: This parameter is not used in Venice image generation but is
                        supported for compatibility with OpenAI API
                    example: natural
                user:
                    type: string
                    description: This parameter is not used in Venice image generation but is
                        supported for compatibility with OpenAI API
                    example: user123
            required:
                - prompt
            additionalProperties: false
...
        CreateEmbeddingRequestSchema:
            type: object
            properties:
                dimensions:
                    type: integer
                    minimum: 1
                    description: The number of dimensions the resulting output embeddings should have.
                encoding_format:
                    type: string
                    enum:
                        - float
                        - base64
                    default: float
                    description: The format to return the embeddings in. Can be either `float` or
                        `base64`.
                    example: float
                input:
                    anyOf:
                        - type: string
                          minLength: 1
                          description: The string that will be turned into an embedding. Cannot be an
                              empty string.
                          example: This is a test.
                          title: string
                        - type: array
                          items:
                              type: string
                          minItems: 1
                          maxItems: 2048
                          description: The array of strings that will be turned into an embedding. Array
                              must be 2048 dimensions or less.
                          example:
                              - This is a test.
                          title: array
                        - type: array
                          items:
                              type: integer
                              minimum: 1
                          minItems: 1
                          maxItems: 2048
                          description: The array of integers that will be turned into an embedding. Array
                              must be 2048 dimensions or less.
                          example:
                              - 1212
                              - 318
                              - 257
                              - 1332
                              - 13
                          title: array
                        - type: array
                          items:
                              type: array
                              items:
                                  type: integer
                              minItems: 1
                          minItems: 1
                          maxItems: 2048
                          description: The array of arrays containing integers that will be turned into an
                              embedding. Array must be 2048 dimensions or less.
                          example:
                              - - 1212
                                - 318
                                - 257
                                - 1332
                                - 13
                          title: array
                    description: Input text to embed, encoded as a string or array of tokens. To
                        embed multiple inputs in a single request, pass an array of strings
                        or array of token arrays. The input must not exceed the max input
                        tokens for the model (8192 tokens), cannot be an empty string, and
                        any array must be 2048 dimensions or less.
                    example: The quick brown fox jumped over the lazy dog
                model:
                    anyOf:
                        - type: string
                        - type: string
                          enum:
                              - text-embedding-bge-m3
                    description: ID of the model to use. You can use the List models API to see all
                        of your available models, or see our Model overview for descriptions
                        of them.
                    example: text-embedding-bge-m3
                user:
                    type: string
                    description: This is an unused parameter and is discarded by Venice. It is
                        supported solely for API compatibility with OpenAI.
            required:
                - input
                - model
            additionalProperties: false
            description: Create embeddings for the supplied input.
            example:
                encoding_format: float
                input: The quick brown fox jumped over the lazy dog
                model: text-embedding-bge-m3
...
paths:
...
    /images/generations:
        post:
            description: Generate an image based on input parameters using an OpenAI
                compatible endpoint. This endpoint does not support the full feature set
                of the Venice Image Generation endpoint, but is compatible with the
                existing OpenAI endpoint.
            operationId: simpleGenerateImage
            parameters:
                - description: Supported compression encodings (gzip, br).
                  in: header
                  name: Accept-Encoding
                  required: false
                  schema:
                      example: gzip, br
                      type: string
            summary: /api/v1/image/generations
            tags:
                - Image
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/SimpleGenerateImageRequest"
            responses:
                "200":
                    description: Successfully generated image
                    headers:
                        Content-Encoding:
                            description: The encoding used to compress the response
                            schema:
                                enum:
                                    - gzip
                                    - br
                                type: string
                        x-venice-is-blurred:
                            description: Indicates if the generated image is blurred. When Safe Venice is
                                enabled, adult material will be returned blurred.
                            required: false
                            schema:
                                type: boolean
                        x-venice-is-content-violation:
                            description: Indicates if the generated image does not meet Venice terms of
                                service.
                            required: false
                            schema:
                                type: boolean
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    created:
                                        type: integer
                                        description: Unix timestamp for when the request was created
                                        example: 1713833628
                                    data:
                                        type: array
                                        items:
                                            anyOf:
                                                - type: object
                                                  properties:
                                                      b64_json:
                                                          type: string
                                                          description: Base64-encoded JSON string of the generated image
                                                          example: iVBORw0KGgoAAAANSUhEUgAA...
                                                  required:
                                                      - b64_json
                                                - type: object
                                                  properties:
                                                      url:
                                                          type: string
                                                          description: Data URL of the generated image
                                                          example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
                                                  required:
                                                      - url
                                required:
                                    - created
                                    - data
                                additionalProperties: false
                "400":
                    description: Invalid request parameters
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DetailedError"
                "401":
                    description: Authentication failed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "402":
                    description: Insufficient USD or Diem balance to complete request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "415":
                    description: Invalid request content-type
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "429":
                    description: Rate limit exceeded
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "500":
                    description: Inference processing failed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "503":
                    description: The model is at capacity. Please try again later.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
...
    /models:
        get:
            description: Returns a list of available models supported by the Venice.ai API
                for both text and image inference.
            operationId: listModels
            security:
                - {}
                - BearerAuth: []
            summary: /api/v1/models
            tags:
                - Models
            parameters:
                - schema:
                      anyOf:
                          - type: string
                            enum:
                                - embedding
                                - image
                                - text
                                - tts
                                - upscale
                                - inpaint
                          - type: string
                            enum:
                                - all
                                - code
                      description: Filter models by type. Use "all" to get all model types.
                      example: text
                  required: false
                  name: type
                  in: query
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/ModelResponse"
                                        description: List of available models
                                    object:
                                        type: string
                                        enum:
                                            - list
                                    type:
                                        anyOf:
                                            - type: string
                                              enum:
                                                  - embedding
                                                  - image
                                                  - text
                                                  - tts
                                                  - upscale
                                                  - inpaint
                                            - type: string
                                              enum:
                                                  - all
                                                  - code
                                        description: Type of models returned.
                                        example: text
                                required:
                                    - data
                                    - object
                                    - type
                "500":
                    description: An unknown error occurred
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
...
    /embeddings:
        post:
            description: Create embeddings for the supplied input.
            operationId: createEmbedding
            parameters:
                - description: Supported compression encodings (gzip, br)
                  in: header
                  name: Accept-Encoding
                  required: false
                  schema:
                      example: gzip, br
                      type: string
            summary: /api/v1/embeddings
            tags:
                - Embeddings
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateEmbeddingRequestSchema"
            responses:
                "200":
                    description: OK
                    headers:
                        Content-Encoding:
                            description: The encoding used to compress the response
                            schema:
                                enum:
                                    - gzip
                                    - br
                                type: string
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                embedding:
                                                    type: array
                                                    items:
                                                        type: number
                                                    description: The embedding vector
                                                index:
                                                    type: integer
                                                    description: The index of this embedding in the list
                                                object:
                                                    type: string
                                                    enum:
                                                        - embedding
                                                    description: The object type, which is always "embedding"
                                            required:
                                                - embedding
                                                - index
                                                - object
                                        description: The list of embeddings generated by the model.
                                    model:
                                        type: string
                                        description: The name of the model used to generate the embedding.
                                    object:
                                        type: string
                                        enum:
                                            - list
                                        description: The object type, which is always "list"
                                    usage:
                                        type: object
                                        properties:
                                            prompt_tokens:
                                                type: integer
                                                description: The number of tokens used by the prompt.
                                            total_tokens:
                                                type: integer
                                                description: The total number of tokens used by the request.
                                        required:
                                            - prompt_tokens
                                            - total_tokens
                                        description: The usage information for the request.
                                required:
                                    - data
                                    - model
                                    - object
                                    - usage
                                example:
                                    data:
                                        - embedding:
                                              - 0.0023064255
                                              - -0.009327292
                                              - 0.015797377
                                          index: 0
                                          object: embedding
                                    model: text-embedding-bge-m3
                                    object: list
                                    usage:
                                        prompt_tokens: 8
                                        total_tokens: 8
                "400":
                    description: Invalid request parameters
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DetailedError"
                "401":
                    description: Authentication failed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "402":
                    description: Insufficient USD or Diem balance to complete request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "415":
                    description: Invalid request content-type
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "429":
                    description: Rate limit exceeded
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "500":
                    description: Inference processing failed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
                "503":
                    description: The model is at capacity. Please try again later.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StandardError"
...